requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==21.2.0
requests
python-dotenv

